requirements.txt
psycopg2==2.9.6
pyTelegramBotAPI==4.10.0
python_dateutil==2.8.2
schedule==1.1.0
